+++ raven/processes/__init__.py
@@ -9,6 +9,7 @@ from .wps_generic_raster_subset import RasterSubsetProcess
 from .wps_generic_terrain_analysis import TerrainAnalysisProcess
 from .wps_generic_zonal_stats import ZonalStatisticsProcess
 from .wps_nalcms_zonal_stats import NALCMSZonalStatisticsProcess
+from .wps_nalcms_zonal_stats_raster import NALCMSZonalStatisticsRasterProcess
 from .wps_ostrich_gr4j_cemaneige import OstrichGR4JCemaNeigeProcess
 from .wps_ostrich_mohyse import OstrichMOHYSEProcess
 from .wps_ostrich_hmets import OstrichHMETSProcess
@@ -58,6 +59,7 @@ geo = [
     TerrainAnalysisProcess(),
     ZonalStatisticsProcess(),
     NALCMSZonalStatisticsProcess(),
+    NALCMSZonalStatisticsRasterProcess(),
 ]
 
 analysis = [