+++ tests/test_wps_ostrich_hbv_ec.py
@@ -88,7 +88,7 @@ class TestOstrichHBVECProcess:
         # checking correctness of NSE (full period 1954-2010 with budget 50 would be NSE=6.034670E-01)
         assert "DIAG_NASH_SUTCLIFFE" in tmp_content[0]
         idx_diag = tmp_content[0].split(",").index("DIAG_NASH_SUTCLIFFE")
-        diag = np.float(tmp_content[1].split(",")[idx_diag])
+        diag = float(tmp_content[1].split(",")[idx_diag])
         np.testing.assert_almost_equal(
             diag, -0.225991, 4, err_msg="NSE is not matching expected value"
         )
@@ -96,7 +96,7 @@ class TestOstrichHBVECProcess:
         # checking correctness of RMSE (full period 1954-2010 would be RMSE=????)
         assert "DIAG_RMSE" in tmp_content[0]
         idx_diag = tmp_content[0].split(",").index("DIAG_RMSE")
-        diag = np.float(tmp_content[1].split(",")[idx_diag])
+        diag = float(tmp_content[1].split(",")[idx_diag])
         np.testing.assert_almost_equal(
             diag, 57.3687, 4, err_msg="RMSE is not matching expected value"
         )